[AdminBundle] Symfony 5 role incompatibility fix
